+++ src/main/java/de/metas/banking/process/BankStatementBasedProcess.java
@@ -9,8 +9,8 @@ import org.compiere.SpringContextHolder;
 import org.compiere.model.I_C_BankStatement;
 import org.compiere.model.I_C_BankStatementLine;
 
-import de.metas.banking.model.BankStatementId;
-import de.metas.banking.model.BankStatementLineId;
+import de.metas.banking.BankStatementId;
+import de.metas.banking.BankStatementLineId;
 import de.metas.banking.payment.IBankStatementPaymentBL;
 import de.metas.banking.service.IBankStatementBL;
 import de.metas.banking.service.IBankStatementDAO;
