+++ src/main/java/de/metas/banking/process/C_BankStatement_AllocatePayment.java
@@ -28,7 +28,7 @@ import org.compiere.model.I_C_BankStatement;
 
 import com.google.common.collect.ImmutableSet;
 
-import de.metas.banking.model.BankStatementId;
+import de.metas.banking.BankStatementId;
 import de.metas.banking.service.IBankStatementDAO;
 import de.metas.document.engine.DocStatus;
 import de.metas.i18n.IMsgBL;
