+++ src/main/java/de/metas/ui/web/bankstatement_reconciliation/BankStatementLineRow.java
@@ -5,7 +5,7 @@ import java.util.Set;
 
 import javax.annotation.Nullable;
 
-import de.metas.banking.model.BankStatementLineId;
+import de.metas.banking.BankStatementLineId;
 import de.metas.currency.Amount;
 import de.metas.ui.web.view.IViewRow;
 import de.metas.ui.web.view.ViewRowFieldNameAndJsonValues;
