+++ src/main/java/de/metas/ui/web/bankstatement_reconciliation/BanksStatementReconciliationViewCreateRequest.java
@@ -4,7 +4,7 @@ import java.util.Collection;
 
 import com.google.common.collect.ImmutableSet;
 
-import de.metas.banking.model.BankStatementLineId;
+import de.metas.banking.BankStatementLineId;
 import de.metas.payment.PaymentId;
 import de.metas.util.Check;
 import lombok.Builder;
