+++ src/test/java/de/metas/ui/web/bankstatement_reconciliation/actions/ReconcilePaymentsCommandTest.java
@@ -30,11 +30,11 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 
 import de.metas.attachments.AttachmentEntryService;
+import de.metas.banking.BankStatementId;
+import de.metas.banking.BankStatementLineId;
+import de.metas.banking.BankStatementLineReference;
+import de.metas.banking.PaySelectionId;
 import de.metas.banking.api.BankAccountId;
-import de.metas.banking.model.BankStatementId;
-import de.metas.banking.model.BankStatementLineId;
-import de.metas.banking.model.BankStatementLineReference;
-import de.metas.banking.model.PaySelectionId;
 import de.metas.banking.model.validator.PaySelectionBankStatementListener;
 import de.metas.banking.payment.IBankStatementPaymentBL;
 import de.metas.banking.payment.IPaySelectionBL;
